Update `ICardInfo` trait to change `game_id` type from `u32` to `u64` for improved compatibility with larger game identifiers.

+++ src/interfaces/cards/info.cairo
@@ -10,5 +10,5 @@ trait ICardInfo<T> {
     ///   - `points (i32)`: The number of points.
     ///   - `multiplier (i32)`: The multiplier applied to the points.
     ///   - `cash (i32)`: The amount of cash available.
-    fn values(self: @T, game_id: u32) -> (i32, i32, i32);
+    fn values(self: @T, game_id: u64) -> (i32, i32, i32);
 }
